backend_web/requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
decorator==5.1.1
Django==4.2.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==23.2
django-redis==5.3.0
django-timezone-field==5.1
djangorestframework==3.14.0
exceptiongroup==1.1.2
executing==1.2.0
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
ipython==8.14.0
jedi==0.19.0
kombu==5.3.1
Markdown==3.4.4
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pytest==7.4.0
pytest-django==4.5.2
pytest-mock==3.11.1
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
